[refactor] 비즈니스 로직 별도의 폴더로 분리
